Refactor: Move PromptFormattingBlock to utilblocks and update imports
- Moved the PromptFormattingBlock implementation from promptformattingblock.py to utilblocks.py, consolidating block definitions.
- Updated the __all__ list in __init__.py to reflect the new location of PromptFormattingBlock.
- Removed the now obsolete promptformattingblock.py file.
